+++ packages/eslint-plugin/src/rules/no-inferrable-types.ts
@@ -125,6 +125,7 @@ export default util.createRule<Options, MessageIds>({
         case AST_NODE_TYPES.TSBooleanKeyword:
           return (
             hasUnaryPrefix(init, '!') ||
+            // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
             isFunctionCall(init, 'Boolean') ||
             isLiteral(init, 'boolean')
           );
@@ -146,6 +147,7 @@ export default util.createRule<Options, MessageIds>({
 
         case AST_NODE_TYPES.TSStringKeyword:
           return (
+            // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
             isFunctionCall(init, 'String') ||
             isLiteral(init, 'string') ||
             init.type === AST_NODE_TYPES.TemplateLiteral