+++ packages/eslint-plugin/src/rules/no-unused-vars-experimental.ts
@@ -141,6 +141,7 @@ export default util.createRule<Options, MessageIds>({
         case ts.SyntaxKind.ImportSpecifier:
         // a namespace import is NOT used, but the default import is used
         case ts.SyntaxKind.NamespaceImport:
+          // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
           report('Import');
           break;
 
@@ -160,6 +161,7 @@ export default util.createRule<Options, MessageIds>({
           break;
 
         case ts.SyntaxKind.PropertyDeclaration:
+          // eslint-disable-next-line @typescript-eslint/internal/prefer-ast-types-enum
           report('Property');
           break;