+++ packages/eslint-plugin/tests/rules/no-floating-promises.test.ts
@@ -2,7 +2,6 @@ import rule from '../../src/rules/no-floating-promises';
 import { RuleTester, getFixturesRootDir } from '../RuleTester';
 
 const rootDir = getFixturesRootDir();
-const messageId = 'floating';
 
 const ruleTester = new RuleTester({
   parserOptions: {
@@ -245,15 +244,39 @@ async function test() {
       errors: [
         {
           line: 3,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 4,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 5,
-          messageId,
+          messageId: 'floating',
+        },
+      ],
+    },
+    {
+      options: [{ ignoreVoid: true }],
+      code: `
+async function test() {
+  Promise.resolve("value");
+}
+`,
+      errors: [
+        {
+          line: 3,
+          messageId: 'floatingVoid',
+          suggestions: [
+            {
+              messageId: 'floatingFixVoid',
+              output: `
+async function test() {
+  void Promise.resolve("value");
+}
+`,
+            },
+          ],
         },
       ],
     },
@@ -268,15 +291,15 @@ async function test() {
       errors: [
         {
           line: 3,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 4,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 5,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -291,15 +314,15 @@ async function test() {
       errors: [
         {
           line: 3,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 4,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 5,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -316,15 +339,15 @@ async function test() {
       errors: [
         {
           line: 5,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 6,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 7,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -338,11 +361,11 @@ async function test() {
       errors: [
         {
           line: 3,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 4,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -357,15 +380,15 @@ async function test() {
       errors: [
         {
           line: 3,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 4,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 5,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -378,7 +401,7 @@ async function test() {
       errors: [
         {
           line: 3,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -392,7 +415,7 @@ async function test() {
       errors: [
         {
           line: 4,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -405,7 +428,7 @@ async function test() {
       errors: [
         {
           line: 3,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -422,15 +445,15 @@ async function test() {
       errors: [
         {
           line: 5,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 6,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 7,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -445,7 +468,7 @@ async function test() {
       errors: [
         {
           line: 5,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -462,15 +485,15 @@ async function test() {
       errors: [
         {
           line: 5,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 6,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 7,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -488,15 +511,15 @@ async function test() {
       errors: [
         {
           line: 6,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 7,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 8,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -517,11 +540,11 @@ async function test() {
       errors: [
         {
           line: 10,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 11,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },
@@ -551,15 +574,15 @@ async function test() {
       errors: [
         {
           line: 18,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 19,
-          messageId,
+          messageId: 'floating',
         },
         {
           line: 20,
-          messageId,
+          messageId: 'floating',
         },
       ],
     },