+++ packages/eslint-plugin/typings/eslint-rules.d.ts
@@ -142,6 +142,22 @@ declare module 'eslint/lib/rules/indent' {
   export = rule;
 }
 
+declare module 'eslint/lib/rules/no-dupe-class-members' {
+  import { TSESLint, TSESTree } from '@typescript-eslint/experimental-utils';
+
+  const rule: TSESLint.RuleModule<
+    'unexpected',
+    [],
+    {
+      Program(): void;
+      ClassBody(): void;
+      'ClassBody:exit'(): void;
+      MethodDefinition(node: TSESTree.MethodDefinition): void;
+    }
+  >;
+  export = rule;
+}
+
 declare module 'eslint/lib/rules/no-dupe-args' {
   import { TSESLint, TSESTree } from '@typescript-eslint/experimental-utils';
 
@@ -431,29 +447,6 @@ declare module 'eslint/lib/rules/no-extra-parens' {
       WhileStatement(node: TSESTree.WhileStatement): void;
       WithStatement(node: TSESTree.WithStatement): void;
       YieldExpression(node: TSESTree.YieldExpression): void;
-    }
-  >;
-  export = rule;
-}
-
-declare module 'eslint/lib/rules/require-await' {
-  import { TSESLint, TSESTree } from '@typescript-eslint/experimental-utils';
-
-  const rule: TSESLint.RuleModule<
-    never,
-    [],
-    {
-      FunctionDeclaration(node: TSESTree.FunctionDeclaration): void;
-      FunctionExpression(node: TSESTree.FunctionExpression): void;
-      ArrowFunctionExpression(node: TSESTree.ArrowFunctionExpression): void;
-      'FunctionDeclaration:exit'(node: TSESTree.FunctionDeclaration): void;
-      'FunctionExpression:exit'(node: TSESTree.FunctionExpression): void;
-      'ArrowFunctionExpression:exit'(
-        node: TSESTree.ArrowFunctionExpression,
-      ): void;
-      ReturnStatement(node: TSESTree.ReturnStatement): void;
-      AwaitExpression(): void;
-      ForOfStatement(node: TSESTree.ForOfStatement): void;
     }
   >;
   export = rule;